refactored rna
